tests: move GenerateRandomMac into the network package
Let's make tests/utils.go shorter.
