+++ tests/network/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "dual_stack_cluster.go",
         "expose.go",
         "framework.go",
+        "mac.go",
         "macvtap.go",
         "networkpolicy.go",
         "port_forward.go",
