+++ tests/network/mac.go
@@ -1,0 +1,16 @@
+package network
+
+import (
+	cryptorand "crypto/rand"
+	"net"
+)
+
+func GenerateRandomMac() (net.HardwareAddr, error) {
+	prefix := net.HardwareAddr{0x02, 0x00, 0x00} // local unicast prefix
+	suffix := make(net.HardwareAddr, 3)
+	_, err := cryptorand.Read(suffix)
+	if err != nil {
+		return nil, err
+	}
+	return append(prefix, suffix...), nil
+}
